Add @beanbag/frontend-buildkit, dependencies for static media building.
This introduces a node.js package called `@beanbag/frontend-buildkit`,
which is a utility package that brings in dependencies for Babel,
Rollup.js, TypeScript, and LessCSS.

Going forward, our products will depend on this package for static media
building. This gives us a versioned set of dependencies that we can
easily reference by any codebase, and in the future a single place for
any plugins or configurations needed.

Testing Done:
Used this package to build static media for Djblets 4 and Review Board 6.

Reviewed at https://reviews.reviewboard.org/r/12774/